faithConnect/backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-dotenv>=0.19.0
pytest>=6.2.5
httpx>=0.18.2
python-multipart>=0.0.5
pydantic>=1.8.2
pytest-asyncio>=0.15.1
pyaudio>=0.2.11
openai>=1.0.0
python-jose[cryptography]>=3.3.0
numpy>=1.24.0
websockets>=10.0
